save all new progress
