+++ beacon-chain/state/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
         "field_trie.go",
         "getters.go",
         "setters.go",
+        "state_trie.go",
         "types.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/state",
